feat: added methods
